i_system.c refactoring, moving all game-specific code to other files
