+++ Source/d_exit.h
@@ -1,0 +1,25 @@
+//
+// Copyright(C) 2022 Fabian Greffrath
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License
+// as published by the Free Software Foundation; either version 2
+// of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// DESCRIPTION:
+//    Exit functions.
+//
+
+#ifndef __D_EXIT__
+#define __D_EXIT__
+
+void I_Quit(void);
+void I_QuitFirst(void);
+void I_QuitLast(void);
+
+#endif
