+++ Source/i_main.c
@@ -37,6 +37,7 @@
 #include "d_main.h"
 #include "i_system.h"
 #include "m_misc2.h"
+#include "d_exit.h"
 
 static void I_SignalHandler(int sig)
 {
@@ -119,66 +120,6 @@ int main(int argc, char **argv)
    D_DoomMain();
    
    return 0;
-}
-
-// Schedule a function to be called when the program exits.
-// If run_if_error is true, the function is called if the exit
-// is due to an error (I_Error)
-// Copyright(C) 2005-2014 Simon Howard
-
-typedef struct atexit_listentry_s atexit_listentry_t;
-
-struct atexit_listentry_s
-{
-    atexit_func_t func;
-    boolean run_on_error;
-    atexit_listentry_t *next;
-    const char *name;
-};
-
-static atexit_listentry_t *exit_funcs[exit_priority_max];
-static exit_priority_t exit_priority;
-
-void I_AtExitPrio(atexit_func_t func, boolean run_on_error,
-                  const char *name, exit_priority_t priority)
-{
-    atexit_listentry_t *entry;
-
-    entry = malloc(sizeof(*entry));
-
-    entry->func = func;
-    entry->run_on_error = run_on_error;
-    entry->next = exit_funcs[priority];
-    entry->name = name;
-    exit_funcs[priority] = entry;
-}
-
-/* I_SafeExit
- * This function is called instead of exit() by functions that might be called
- * during the exit process (i.e. after exit() has already been called)
- */
-
-void I_SafeExit(int rc)
-{
-  atexit_listentry_t *entry;
-
-  // Run through all exit functions
-
-  for (; exit_priority < exit_priority_max; ++exit_priority)
-  {
-    while ((entry = exit_funcs[exit_priority]))
-    {
-      exit_funcs[exit_priority] = exit_funcs[exit_priority]->next;
-
-      if (rc == 0 || entry->run_on_error)
-      {
-//      fprintf(stderr, "Exit Sequence[%d]: %s (%d)\n", exit_priority, entry->name, rc);
-        entry->func();
-      }
-    }
-  }
-
-  exit(rc);
 }
 
 //----------------------------------------------------------------------------
